// packages/utils/src/runtime-schema.ts
export interface FieldSpec {
  name: string;
  type: string;
  isArray: boolean;
  optional: boolean;
}
export interface NodeSpec {
  name: string;
  isNode: boolean;
  fields: FieldSpec[];
}